// frontend/src/api/generated/model/dtoMemberResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Diagram Backend API
 * Thin persistence API for diagram streams, schema streams, and authenticated member metadata.
 * OpenAPI spec version: 1.0
 */

export interface DtoMemberResponse {
  createdAt?: string;
  displayName?: string;
  email?: string;
  id?: string;
  lastSeenAt?: string;
  organizationId?: string;
  principalId?: string;
  profilePictureUrl?: string;
  updatedAt?: string;
}